third_party/WebKit/LayoutTests/platform/android/http/tests/origin_trials/webexposed/animationworklet-origin-trial-interfaces-worklet-scope-expected.txt

CONSOLE MESSAGE: line 13: List all interfaces in the animation worklet global scope.
CONSOLE MESSAGE: line 151: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface CountQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface EffectProxy
CONSOLE MESSAGE: line 151:     getter localTime
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter localTime
CONSOLE MESSAGE: line 151: interface ReadableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method cancel
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getReader
CONSOLE MESSAGE: line 151:     method pipeThrough
CONSOLE MESSAGE: line 151:     method pipeTo
CONSOLE MESSAGE: line 151:     method tee
CONSOLE MESSAGE: line 151: interface TransformStream
CONSOLE MESSAGE: line 151:     getter readable
CONSOLE MESSAGE: line 151:     getter writable
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 153: interface WorkletGroupEffectProxy
CONSOLE MESSAGE: line 153:     method constructor
CONSOLE MESSAGE: line 153:     method getChildren
CONSOLE MESSAGE: line 151: interface WritableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method abort
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getWriter
CONSOLE MESSAGE: line 151: global object
CONSOLE MESSAGE: line 151:     attribute console
CONSOLE MESSAGE: line 151:     method gc
CONSOLE MESSAGE: line 151:     method registerAnimator
